логгер и дебаггер теперь расположены отдельно от app.py, loader.py. Исправлена инициализ. статики сайта
