+++ .gitignore
@@ -109,7 +109,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
-.env
+data/.env
 .venv
 env/
 venv/
