+++ django_admin/django_admin/settings.py
@@ -12,22 +12,19 @@ https://docs.djangoproject.com/en/3.2/ref/settings/
 from pathlib import Path
 import os
 
+from data import config
+
 # Build paths inside the project like this: BASE_DIR / 'subdir'.
 
-# import _locale
-# _locale._getdefaultlocale = (lambda *args: ['en_US', 'utf8'])
 
-# * root path of django project
 BASE_DIR = Path(__file__).resolve().parent.parent
-
-# * root path of project
 ROOT_DIR = os.getcwd()
 
 # Quick-start development settings - unsuitable for production
 # See https://docs.djangoproject.com/en/3.2/howto/deployment/checklist/
 
 # SECURITY WARNING: keep the secret key used in production secret!
-SECRET_KEY = "django-insecure-g2vvo@o5pl(r6ee%0*p$#!iyu(@#vv)_@!-71i884_#v!9-45n"
+SECRET_KEY = config.SECRET_KEY_DJ#"django-insecure-bmp4k!f5lh=j19=u)z#-^gu_=jy-zxsh_#xfvg_!e7=aj=rril"
 
 # SECURITY WARNING: don't run with debug turned on in production!
 DEBUG = True
@@ -36,6 +33,7 @@ ALLOWED_HOSTS = ['*']
 
 
 # Application definition
+
 INSTALLED_APPS = [
     'django_admin.bot.apps.BotConfig',
     'django_admin.feedback.apps.FeedbackConfig',
@@ -87,11 +85,11 @@ WSGI_APPLICATION = 'django_admin.django_admin.wsgi.application'
 DATABASES = {
     'default': {
         'ENGINE': 'django.db.backends.postgresql_psycopg2',
-        'NAME': 'chat-bot-data',
-        'USER': 'postgres',
-        'PASSWORD': '860269',
-        'HOST': '127.0.0.1',
-        'PORT': '5432'
+        'NAME': config.DB_NAME,
+        'USER': config.DB_USER,
+        'PASSWORD': config.DB_PASS,
+        'HOST': config.DB_HOST,
+        'PORT': config.DB_PORT
     }
 }
 
@@ -135,14 +133,13 @@ USE_TZ = True
 MEDIA_ROOT = os.path.join(BASE_DIR, "django_admin/uploads")
 MEDIA_URL = '/uploads/'
 
-STATIC_URL = '/static/'
-
+STATIC_URL = 'static/'
 # Для локального сервера
-STATICFILES_DIRS = [
-   os.path.join(BASE_DIR, 'static')
-]
+# STATICFILES_DIRS = [
+#     os.path.join(BASE_DIR, 'static')
+# ]
 
-# STATIC_ROOT = os.path.join(BASE_DIR, '/static')
+STATIC_ROOT = os.path.join(BASE_DIR, STATIC_URL)
 
 # Default primary key field type
 # https://docs.djangoproject.com/en/3.2/ref/settings/#default-auto-field
