+++ django_admin/django_admin/urls.py
@@ -16,9 +16,10 @@ Including another URLconf
 from django.contrib import admin
 
 from django.urls import path, include
-from django_admin.bot import views
+from django.conf.urls.static import static
+from django_admin.django_admin import settings
 
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include("django_admin.bot.urls"))
-]
+] + static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
