+++ handlers/groups/main_menu_group.py
@@ -7,7 +7,8 @@ from django.db.models import QuerySet
 from django_admin.bot.models import Help_content
 from django_admin.service.models import Users, ChatIDAdmission, ChatIDDirections
 from filters import IsGroup
-from loader import dp, bot, debugger
+from loader import dp, bot
+from utils.debugger import debugger
 from aiogram.types import Message, CallbackQuery
 
 from keyboards.inline import buttons as btn
